Move generic utilities into pkg
