+++ internal/caveats/diff.go
@@ -3,9 +3,9 @@ package caveats
 import (
 	"golang.org/x/exp/maps"
 
-	"github.com/authzed/spicedb/internal/util"
 	"github.com/authzed/spicedb/pkg/caveats/types"
 	core "github.com/authzed/spicedb/pkg/proto/core/v1"
+	"github.com/authzed/spicedb/pkg/util"
 )
 
 // DeltaType defines the type of caveat deltas.
